Got KD influence on Morale Missions setup fully

+++ Scripts/Story/Morale_System.lua
@@ -24,10 +24,10 @@ function Definitions()
     }
 
     morale_event_table = {
-        ["Morale_Lost_Battle"] = {Value = 1, Subtract = true}, -- left over from figuring out why code isnt working, too lazy to change even tho it doenst make sense, but i dont wanna change stuff
-        ["Morale_Lost_Battle_Major"] = {Value = 3, Subtract = true},
-        ["Morale_Won_Battle"] = {Value = 1, Subtract = false},
-        ["Morale_Won_Battle_Major"] = {Value = 3, Subtract = false},
+        ["Morale_Lost_Battle"] = {Value = 1, Subtract = true, KD_Influence = true},
+        ["Morale_Lost_Battle_Major"] = {Value = 3, Subtract = true, KD_Influence = true},
+        ["Morale_Won_Battle"] = {Value = 1, Subtract = false, KD_Influence = true},
+        ["Morale_Won_Battle_Major"] = {Value = 3, Subtract = false, KD_Influence = true},
         ["Morale_Construction_Event_Minor"] = {Value = 1, Subtract = false},
         ["Morale_Construction_Event"] = {Value = 2, Subtract = false},
         ["Morale_Construction_Event_Major"] = {Value = 3, Subtract = false},
@@ -35,7 +35,9 @@ function Definitions()
         ["Hero_Killed"] = {Value = 3, Subtract = false},
     }
 
-    UNSC_Kill_Ratio_Table = {1, } -- the index is the morale gain from the kill ratio at that index
+    UNSC_Kill_Ratio_Table = {0.0625, 0.07, 0.1 , 0.16, 0.3} -- the index is the morale gain from the kill ratio at that index
+
+    COVN_Kill_Ratio_Table = {0.3, 0.5, 1, 2, 3}
 
     hero_status_table = {
         UNSC_POA = {Current_Status = false, Equation = "Is_POA_Alive", Object = nil, Owner = nil},
@@ -65,7 +67,6 @@ function Init_Morale_System(message)
         for hero, status in pairs(hero_status_table) do
             if EvaluatePerception(status.Equation, player) == 1 then
                 hero_status_table[hero].Current_Status = true
-                hero_status_table[hero].Last_Status = true
                 
                 local hero_object = Find_First_Object(hero)
 
@@ -142,7 +143,7 @@ function Check_Hero_Status()
 
                 hero_status_table[hero].Current_Status = true
 
-                if status.Object == nil then
+                if status.Object == nil or (not TestValid(status.Object)) then
 
                     local hero_object = Find_First_Object(hero)
 
@@ -178,12 +179,10 @@ function Modify_Morale(event_table)
 
     DebugMessage("%s -- Event Morale Value: %s, Subtract: %s", tostring(Script), tostring(Morale_Value), tostring(bad))
 
-    local Next_Morale_Level = 0
+    local Next_Morale_Level = global_morale_level + Morale_Value
 
     if bad then
         Next_Morale_Level = global_morale_level - Morale_Value
-    else
-        Next_Morale_Level = global_morale_level + Morale_Value
     end
 
     DebugMessage("%s -- Next Morale Value: %s", tostring(Script), tostring(Next_Morale_Level))
@@ -210,22 +209,73 @@ function Get_Morale_Influence()
     PrintTable(Morale_Values)
 
     if type(Morale_Values) == "table" then
-        return Morale_Values
+        if Morale_Values.KD_Influence == true then
+            local New_Morale_Value = Morale_Kill_Ratio_Influence(Morale_Values.Value, Morale_Values.Subtract)
+
+            return {Value = New_Morale_Value, Subtract = Morale_Values.Subtract}
+        else
+            return Morale_Values
+        end
     else
         return {Value = 0, Subtract = false}
     end
 end
 
-function Morale_Kill_Ratio_Influence()
-    local Base_Morale = Get_Morale_Influence()
+function Morale_Kill_Ratio_Influence(Base_Morale, is_loss)
+
+    if is_loss ~= true then
+        is_loss = false
+    end
+
+    if type(Base_Morale) ~= "number" then
+        return 0
+    end
 
     local Kill_Ratio = GlobalValue.Get("Morale_Kill_Ratio")
+
+    DebugMessage("%s -- Kill Ratio: %s", tostring(Script), tostring(Kill_Ratio))
+
+    if Kill_Ratio == nil then
+        return Base_Morale
+    end
 
     if Kill_Ratio <= 0 then -- if this is true we didnt get the proper kill ratio
         return Base_Morale
     end
 
+    local Nearest_Kill_Ratio_Morale_Gain = 0
+
+    local Kill_Ratio_Table = UNSC_Kill_Ratio_Table
+
+    if player == Find_Player("EMPIRE") then
+        Kill_Ratio_Table = COVN_Kill_Ratio_Table
+    end
+
+    for morale_gain, ratio in pairs(Kill_Ratio_Table) do
+        if Kill_Ratio >= ratio then
+            Nearest_Kill_Ratio_Morale_Gain = morale_gain
+        end
+    end
+
+    DebugMessage("%s -- Morale Gain for KD %s: %s", tostring(Script), tostring(Kill_Ratio), tostring(Nearest_Kill_Ratio_Morale_Gain))
+
+    local Final_Morale_Gain = Nearest_Kill_Ratio_Morale_Gain
+
+    local Max_Morale_Gain = tableLength(Kill_Ratio_Table)
+
+    if is_loss then
+        Final_Morale_Gain = Max_Morale_Gain - Nearest_Kill_Ratio_Morale_Gain
+    end
     
+    if Final_Morale_Gain > Max_Morale_Gain then
+        Final_Morale_Gain = Max_Morale_Gain
+    elseif Final_Morale_Gain < 0 then
+        Final_Morale_Gain = 0
+    end
+
+    DebugMessage("%s -- Final Morale Gain: %s", tostring(Script), tostring(Final_Morale_Gain))
+
+    return Final_Morale_Gain
 end
 
 function Flush(message)
